enum values with underlying type 'NSInteger' should not be used as format arguments; add an explicit cast to 'long' instead

+++ Iterable-iOS-SDK/IterableAPI.m
@@ -386,7 +386,7 @@ NSString * const endpoint = @"https://api.iterable.com/api/";
             result = @"APNS_SANDBOX";
             break;
         default:
-            NSLog(@"Unexpected PushServicePlatform: %d", pushServicePlatform);
+            NSLog(@"Unexpected PushServicePlatform: %d", (long)pushServicePlatform);
     }
 
     return result;
